// src/public-api/model/orderDto.ts
/**
 * MISCore Shop API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { OrderItemDto } from './orderItemDto';
import { AddressDto } from './addressDto';
import { PaymentDto } from './paymentDto';


export interface OrderDto { 
    id?: number;
    orderNumber: string;
    userId: string;
    userName?: string | null;
    email?: string | null;
    orderStatusId?: number;
    orderStatusName?: string | null;
    orderStatus?: string | null;
    paymentMethodId?: number | null;
    paymentMethodName?: string | null;
    shippingMethodId?: number | null;
    shippingMethodName?: string | null;
    subtotal?: number;
    taxAmount?: number;
    shippingCost?: number;
    discountAmount?: number;
    totalAmount?: number;
    billingAddress?: AddressDto;
    shippingAddress?: AddressDto;
    notes?: string | null;
    trackingNumber?: string | null;
    orderDate?: string;
    shippedDate?: string | null;
    deliveredDate?: string | null;
    orderItems?: Array<OrderItemDto> | null;
    payments?: Array<PaymentDto> | null;
}